XMDA testing
This commit is for testing the XMDA engine driver and interfaces. The
xdma-test tool can be used for this purpose, and has some documentation on
setting up the test environment on the host side:

https://github.com/eddiejames/xdma-test

w5 host is currently set up to easily test. No need to compile the
driver; the module file (ast-bmc-pcie.ko) is found in /home/root/xdma/
so just follow the directions in xdma-test/README.md to remove the
existing ast driver and probe my driver.

wsbmc022 host is also set up, but you have to choose to boot from the
4.1X kernel, NOT the 5.0 kernel in petitboot. Module file should be in
/home/root/pcie-xdma/

Change-Id: Ibf6d888b3964bdb63cd5e1f65a2e78c525a8272e
Signed-off-by: Eddie James <eajames@us.ibm.com>

Conflicts:
	meta-aspeed/recipes-kernel/linux/ast2500/defconfig